// src/lib/enums.ts
export enum UserStatus {
  ACTIVE = 'active',
  RESTRICTED = 'restricted',
  BANNED = 'banned',
}

export enum Gender {
  MALE = 'male',
  FEMALE = 'female',
  OTHER = 'other',
}

export enum Role {
  ADMIN = 'admin',
  MODERATOR = 'moderator',
  MEMBER = 'member',
}

export enum ThreadStatus {
  ACTIVE = 'active',
  LOCKED = 'locked',
  HIDDEN = 'hidden',
}